[CAN-25] refactor: 노트 리스트 컴포넌트 분리
